+++ system/bootstrap.php
@@ -6,7 +6,7 @@
  * Allows for some configuration in the cs-config.php
  *
  * @package Cornerstone
- * @version 0.2.18
+ * @version 0.2.19
  * @author Damien Peden <support@dpdesignz.co>
  * @copyright 2019-2020 dpDesignz. All Rights Reserved
  * @license This script can be used FREE of charge for any commercial or personal projects. Enjoy!